Change delta time to float to allow for fractions of seconds

+++ include/GameObject.hpp
@@ -8,7 +8,7 @@ public:
     GameObject(SDL_Renderer *renderer, std::string name, int x, int y, SDL_Color color);
     ~GameObject();
 
-    void update(unsigned int deltaTime, int worldWidth, int worldHeight, int gravity);
+    void update(float deltaTime, int worldWidth, int worldHeight, int gravity);
     void render();
     int getX() const { return x; }
     int getY() const { return y; }
@@ -23,7 +23,7 @@ public:
     SDL_Color color;
 
 private:
-    void applyGravity(unsigned int deltaTime, int gravity, int worldHeight);
+    void applyGravity(float deltaTime, int gravity, int worldHeight);
 
     SDL_Renderer *renderer;
     std::string name;
